+++ generate-client.sh
@@ -1,0 +1,13 @@
+#!/bin/bash
+# TODO: Move client generation to docker image
+
+export AdvanceConfigurationPath=./config.yml
+export Serilog__MinimumLevel=Verbose
+export ASPNETCORE_ENVIRONMENT=Development
+export ASPNETCORE_SUPPRESSSTATUSMESSAGES=true
+export ELASTICSEARCH_URL_CSV=http://system-events-elasticsearch:9200/
+export ELASTICSEARCH_INDEX=sysevents
+export ELASTICSEARCH_TIMEOUT_MS=5000
+export ELASTICSEARCH_DATETIME_FORMAT=yyyy-MM-dd'T'HH:mm:ssZ
+
+dotnet build system-events/clients/dotnet --configuration Release