+++ slack-app-backend/src/Configuration/IAppConfiguration.cs
@@ -1,0 +1,17 @@
+using System;
+
+namespace SlackAppBackend.Configuration
+{
+    public interface IAppConfiguration
+    {
+        Uri SystemEventServiceUri { get; }
+
+        string SlackBotUserOAuthAccessToken { get; }
+
+        string SlackSignedSecret { get; }
+
+        bool ShowCustomCategory { get; }
+
+        bool ShowPredefinedCategory { get; }
+    }
+}