+++ slack-app-backend/src/Enums/SlackCommand.cs
@@ -1,0 +1,8 @@
+namespace SlackAppBackend.Enums
+{
+    public enum SlackCommand
+    {
+        Create,
+        End
+    }
+}