+++ slack-app-backend/src/Models/OpenSlackModalRequest.cs
@@ -1,0 +1,8 @@
+namespace SlackAppBackend.Models.Slack
+{
+    public class OpenSlackModalRequest
+    {
+        public string trigger_id { get; set; }
+        public string view { get; set; }
+    }
+}