+++ slack-app-backend/src/Utils/Interfaces/IMonitoredSystemEventServiceClient.cs
@@ -1,0 +1,13 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using SystemEvents.Api.Client.CSharp.Contracts;
+
+namespace SlackAppBackend.Utils.Interfaces
+{
+    public interface IMonitoredSystemEventServiceClient
+    {
+         Task<EventResponse> EventPostAsync(EventRequestModel model);
+         Task EventEndPostAsync(string eventId);
+         Task<ICollection<Category>> CategoryAllGetAsync();
+    }
+}