+++ system-events/src/Configuration/IElasticsearchClientConfiguration.cs
@@ -2,7 +2,7 @@ namespace SystemEvents.Configuration
 {
     public interface IElasticsearchClientConfiguration
     {
-         public string UrlCsv { get; }
+        public string UrlCsv { get; }
 
         public string DefaultIndex { get; }